requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.108.0
ffmpeg==1.4
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
idna==3.6
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
openai-whisper==20231117
PyAudio==0.2.14
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
python-multipart==0.0.6
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
tiktoken==0.5.2
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
websockets==12.0
